Add superagent as a replacement for http for requests
